requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
certifi==2026.2.25
charset-normalizer==3.4.7
distro==1.9.0
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.14.0
jsonpatch==1.33
jsonpointer==3.1.1
langchain==0.2.10
langchain-core==0.2.43
langchain-openai==0.1.25
langchain-text-splitters==0.2.4
langgraph==1.1.7
langgraph-checkpoint==4.0.2
langgraph-prebuilt==1.0.9
langgraph-sdk==0.3.13
langsmith==0.1.147
multidict==6.7.1
numpy==1.26.4
openai==1.109.1
orjson==3.11.8
ormsgpack==1.12.2
packaging==24.2
propcache==0.4.1
pydantic==2.13.2
pydantic_core==2.46.2
python-dotenv==1.2.2
PyYAML==6.0.3
regex==2026.4.4
requests==2.33.1
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.49
tenacity==8.5.0
tiktoken==0.12.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
xxhash==3.6.0
yarl==1.23.0
zstandard==0.25.0
langchain==0.2.10
langchain-openai
openai
python-dotenv
fastapi
uvicorn
